+++ frontend/src/app/workspace/workspace-header/workspace-header.component.ts
@@ -23,7 +23,6 @@ export class WorkspaceHeaderComponent {
   @Output() tabAdd = new EventEmitter<void>();
   @Output() tabRemove = new EventEmitter<string>();
   @Output() tabRenameFinish = new EventEmitter<{ id: string; name: string }>();
-  @Output() openBackground = new EventEmitter<void>();
   @Output() openWidgetSelector = new EventEmitter<void>();
   @Output() resetWorkspace = new EventEmitter<void>();
   @Output() save = new EventEmitter<void>();